fix(wasm): fix js.Value.Int64() and self-capturing closure build errors
- Replace `seq.Int64()` with `int64(seq.Int())` since `js.Value` has no
  `Int64()` method in the `syscall/js` package (WASM target)
- Lift `thenHandler`, `jsonThenHandler`, `catchHandler` `js.Func` variables
  to the outer scope so the `defer *.Release()` inside each closure can
  reference them without capturing an uninitialized variable
- Update package-lock.json

+++ cmd/wasm/opfs_replica.go
@@ -611,7 +611,7 @@ func getPersistedSeqNumbers() map[string]int64 {
 		val := opfsMeta.Get(key)
 		if seq := val.Get("browserSeq"); seq.Truthy() {
 			if seq.Type() == js.TypeNumber {
-				seqs[key] = seq.Int64()
+				seqs[key] = int64(seq.Int())
 			}
 		}
 	}
